feat(系统设置): 认证登陆调试首次跳转调试中
